# apps/web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "AuthSessionResponse": {
        "properties": {
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "authorization": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorizationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "csrf_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csrf Token"
          },
          "discord_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discord User Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "authenticated"
        ],
        "title": "AuthSessionResponse",
        "type": "object"
      },
      "AuthorizationResponse": {
        "properties": {
          "capabilities": {
            "items": {
              "enum": [
                "app.view",
                "players.manage",
                "coaching.configure",
                "usage.view",
                "admin.manage_rbac"
              ],
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "community": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community"
          },
          "is_member": {
            "default": false,
            "title": "Is Member",
            "type": "boolean"
          },
          "role_ids": {
            "items": {
              "type": "string"
            },
            "title": "Role Ids",
            "type": "array"
          },
          "status": {
            "default": "unavailable",
            "enum": [
              "member",
              "not_member",
              "unavailable",
              "unknown"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "title": "AuthorizationResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "api_version"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "LogoutResponse": {
        "properties": {
          "signed_out": {
            "title": "Signed Out",
            "type": "boolean"
          }
        },
        "required": [
          "signed_out"
        ],
        "title": "LogoutResponse",
        "type": "object"
      },
      "RbacMappingResponse": {
        "properties": {
          "capability": {
            "enum": [
              "app.view",
              "players.manage",
              "coaching.configure",
              "usage.view",
              "admin.manage_rbac"
            ],
            "title": "Capability",
            "type": "string"
          },
          "community": {
            "title": "Community",
            "type": "string"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "string"
          },
          "role_id": {
            "title": "Role Id",
            "type": "string"
          }
        },
        "required": [
          "community",
          "guild_id",
          "role_id",
          "capability"
        ],
        "title": "RbacMappingResponse",
        "type": "object"
      },
      "RbacMappingsResponse": {
        "properties": {
          "mappings": {
            "items": {
              "$ref": "#/components/schemas/RbacMappingResponse"
            },
            "title": "Mappings",
            "type": "array"
          }
        },
        "required": [
          "mappings"
        ],
        "title": "RbacMappingsResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Versioned HTTP boundary for the Group Raiding applications.",
    "title": "Group Raiding API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/discord/callback": {
      "get": {
        "operationId": "discordCallback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          },
          {
            "in": "cookie",
            "name": "group_raiding_oauth_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Raiding Oauth State"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discord Callback",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/discord/login": {
      "get": {
        "operationId": "loginWithDiscord",
        "parameters": [
          {
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discord Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "group_raiding_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Raiding Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "operationId": "getAuthSession",
        "parameters": [
          {
            "in": "cookie",
            "name": "group_raiding_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Raiding Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/session/refresh": {
      "post": {
        "operationId": "refreshAuthSession",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "group_raiding_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Raiding Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/healthz": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/rbac/mappings": {
      "get": {
        "operationId": "listRbacMappings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RbacMappingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Rbac Mappings",
        "tags": [
          "authorization"
        ]
      }
    }
  }
}